Add basic support for Boot templates #1
This also cleans up a fair bit of the existing Leiningen code to be a
bit less Leiningen-specific and a bit more Boot-specific.

The new boot.new.templates namespace no longer relies on Leiningen code
at all. Not all of the cleanup from that has been back-ported to
leiningen.new.templates but it probably will be.

There's still a bunch of Leiningen repository and classpath code that
needs to be converted to Boot specific stuff (for example it still
consults Leiningen's user profiles!).