Remove dependence on Leiningen #1
Neither Boot new, nor either of leiningen.new.templates or
boot.new.templates relies on Leiningen any more (although Boot new still
relies on Bultitude at the moment).

Leiningen templates themselves may still rely on leiningen-core so that
is still a dependency (we could probably defer merging that until we've
established that we're working with a Leiningen template).

Boot templates are "supported" but not yet tested so I'm not closing
this out yet.